Move BeamElement into beam directory and update imports accordingly
